Share handoff lifecycle helpers
